+++ LuiExtended/modules/ChatAnnouncements/ChatAnnouncementsConstants.lua
@@ -9,6 +9,8 @@ local LUIE = LUIE
 -- ChatAnnouncements namespace
 --- @class (partial) ChatAnnouncements
 --- @field SV CADefaults
+--- @field ChatOutput LUIE_ChatOutput
+--- @field ChatOutputClass LUIE_ChatOutput
 local ChatAnnouncements = {}
 
 --- @class (partial) ChatAnnouncements